Move initialization to the module factory also.
